requirements.txt
openai~=0.27.4
PyYAML~=6.0